Improve address handling
This commit adds support for parsing the patron addresses into their
parts. Addresses are stored as a single text field in Alice, but it
tries to parse the field into separate fields on export. Unfortunately,
this doesn't work very well, especially with the German-style addresses
we store at the IELD. So, the script handles several interpretations of
the address field by the Alice exporter.